USB Module changes
Generic USB in parts of the 4.18 kernel and mainly 4.19 kernel have been changed from being included in the kernel to modular.
If the kernel is being built with mkinitscript for the initramfs you will have no usb support with or without the surface patches period. I have changed these to y to force them to be included in the kernel so that they are picked up by initramfs. This but is mainly in Yocto not sure about other places.