Dynamically detect the packaging system of the distro
Since the official packages are offered as .deb packages,
it doesn't make much sense to offer installing them on a
pacman-based, or rpm-based distribution. Instead, the
package manager is detected dynamically (i.e. if dpkg is
found you can install .deb, if pacman is found you can
install .pkg.tar.xz etc.), and the apropreate source for
kernel / libwacom packages is linked.

It might be possible to install them automatically as well,
but some distributions might require custom setup so it is
probably safer to only offer autoinstall for the upstream
packages.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>